+++ v8go.cc
@@ -411,74 +411,6 @@ RtnValue FunctionTemplateGetFunction(TemplatePtr ptr, ContextPtr ctx) {
 
 /********** Context **********/
 
-ContextPtr NewContext(IsolatePtr iso,
-                      TemplatePtr global_template_ptr,
-                      int ref) {
-  Locker locker(iso);
-  Isolate::Scope isolate_scope(iso);
-  HandleScope handle_scope(iso);
-
-  Local<ObjectTemplate> global_template;
-  if (global_template_ptr != nullptr) {
-    global_template = global_template_ptr->ptr.Get(iso).As<ObjectTemplate>();
-  } else {
-    global_template = ObjectTemplate::New(iso);
-  }
-
-  // For function callbacks we need a reference to the context, but because of
-  // the complexities of C -> Go function pointers, we store a reference to the
-  // context as a simple integer identifier; this can then be used on the Go
-  // side to lookup the context in the context registry. We use slot 1 as slot 0
-  // has special meaning for the Chrome debugger.
-  Local<Context> local_ctx = Context::New(iso, nullptr, global_template);
-  local_ctx->SetEmbedderData(1, Integer::New(iso, ref));
-
-  m_ctx* ctx = new m_ctx;
-  ctx->ptr.Reset(iso, local_ctx);
-  ctx->iso = iso;
-  return ctx;
-}
-
-int ContextRetainedValueCount(ContextPtr ctx) {
-  return ctx->vals.size();
-}
-
-RtnValue RunScript(ContextPtr ctx, const char* source, const char* origin) {
-  LOCAL_CONTEXT(ctx);
-
-  RtnValue rtn = {};
-
-  MaybeLocal<String> maybeSrc =
-      String::NewFromUtf8(iso, source, NewStringType::kNormal);
-  MaybeLocal<String> maybeOgn =
-      String::NewFromUtf8(iso, origin, NewStringType::kNormal);
-  Local<String> src, ogn;
-  if (!maybeSrc.ToLocal(&src) || !maybeOgn.ToLocal(&ogn)) {
-    rtn.error = ExceptionError(try_catch, iso, local_ctx);
-    return rtn;
-  }
-
-  ScriptOrigin script_origin(ogn);
-  Local<Script> script;
-  if (!Script::Compile(local_ctx, src, &script_origin).ToLocal(&script)) {
-    rtn.error = ExceptionError(try_catch, iso, local_ctx);
-    return rtn;
-  }
-  Local<Value> result;
-  if (!script->Run(local_ctx).ToLocal(&result)) {
-    rtn.error = ExceptionError(try_catch, iso, local_ctx);
-    return rtn;
-  }
-  m_value* val = new m_value;
-  val->id = 0;
-  val->iso = iso;
-  val->ctx = ctx;
-  val->ptr = Global<Value>(iso, result);
-
-  rtn.value = tracked_value(ctx, val);
-  return rtn;
-}
-
 RtnValue JSONParse(ContextPtr ctx, const char* str) {
   LOCAL_CONTEXT(ctx);
   RtnValue rtn = {};
@@ -546,18 +478,6 @@ void ValueRelease(ValuePtr ptr) {
   ptr->ctx->vals.erase(ptr->id);
   ptr->ptr.Reset();
   delete ptr;
-}
-
-ValuePtr ContextGlobal(ContextPtr ctx) {
-  LOCAL_CONTEXT(ctx);
-  m_value* val = new m_value;
-  val->id = 0;
-
-  val->iso = iso;
-  val->ctx = ctx;
-  val->ptr = Global<Value>(iso, local_ctx->Global());
-
-  return tracked_value(ctx, val);
 }
 
 /********** Value **********/
